fixed admin config

+++ security/src/main/java/ee/gaile/service/security/settings/AdminConfig.java
@@ -2,9 +2,9 @@ package ee.gaile.service.security.settings;
 
 import ee.gaile.entity.users.Users;
 import ee.gaile.enums.EnumRoles;
+import ee.gaile.service.security.SecurityConfig;
 import ee.gaile.service.security.UserRepository;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Component;
 
 import javax.annotation.PostConstruct;
@@ -12,11 +12,11 @@ import javax.annotation.PostConstruct;
 @Component
 public class AdminConfig {
     private final UserRepository userRepository;
-    private final PasswordEncoder encoder;
+    private final SecurityConfig securityConfig;
 
-    private AdminConfig(UserRepository userRepository, PasswordEncoder encoder) {
+    private AdminConfig(UserRepository userRepository, SecurityConfig securityConfig) {
         this.userRepository = userRepository;
-        this.encoder = encoder;
+        this.securityConfig = securityConfig;
     }
 
     @Value("${cv.admin}")
@@ -31,11 +31,16 @@ public class AdminConfig {
     @PostConstruct
     private void changeAdmin() {
         Users admin = userRepository.findByRole(EnumRoles.ROLE_ADMIN);
-        if (admin != null) {
-            userRepository.delete(admin);
+        if (admin == null || !admin.getUsername().equals(adminName) || !admin.getEmail().equals(email)
+                || !securityConfig.passwordEncoder()
+                .matches(password, admin.getPassword())) {
+
+            if (admin != null) {
+                userRepository.delete(admin);
+            }
+            Users user = new Users(adminName, email, securityConfig.passwordEncoder().encode(password), EnumRoles.ROLE_ADMIN);
+            userRepository.save(user);
         }
-        Users user = new Users(adminName, email, encoder.encode(password), EnumRoles.ROLE_ADMIN);
-        userRepository.save(user);
     }
 
 }
